# card-service/docs/openapi-snapshot.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:55961
  description: Generated server url
paths:
...
  /card-api/v1/cards:
    get:
      tags:
      - card-api-controller
      operationId: listCards
      parameters:
      - name: customerId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListCard"
    post:
      tags:
      - card-api-controller
      operationId: issueCard
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCardRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCard"
...
  /api/v1/simulate/balance:
    post:
      tags:
      - terminal-simulator-controller
      operationId: balance
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseSimulateResponse"
...
  /api/v1/cards:
    get:
      tags:
      - card-controller
      operationId: listCards_1
      parameters:
      - name: customerId
        in: query
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListCard"
    post:
      tags:
      - card-controller
      operationId: issueCard_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCardRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCard"
  /api/v1/cards/{id}:
    get:
      tags:
      - card-controller
      operationId: getCard
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCard"
    post:
      tags:
      - card-controller
      operationId: command
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: command
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCard"
...
  /card-api/v1/cards/{id}:
    get:
      tags:
      - card-api-controller
      operationId: getCard_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseCard"
...
  /api/v1/cards/{id}/balance:
    get:
      tags:
      - card-controller
      operationId: getBalance
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseBalanceResult"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        field:
          type: string
...
    Card:
      type: object
      properties:
        id:
          type: string
          format: uuid
        panEncrypted:
          type: string
        panHash:
          type: string
        panPrefix:
          type: string
        panSuffix:
          type: string
        expiryDate:
          type: string
        cvvEncrypted:
          type: string
        cardSequenceNo:
          type: integer
          format: int32
        cardType:
          type: string
          enum:
          - DEBIT
          - PREPAID
          - CREDIT
        status:
          type: string
          enum:
          - ISSUED
          - ACTIVE
          - BLOCKED
          - EXPIRED
          - CANCELLED
          - ORDERED
          - PRODUCED
          - DISPATCHED
          - ACTIVATION_PENDING
        virtualFlag:
          type: boolean
        customerId:
          type: string
          format: uuid
        linkedEntityId:
          type: string
          format: uuid
        product:
          $ref: "#/components/schemas/CardProduct"
        pinRetryCount:
          type: integer
          format: int32
        pinSet:
          type: boolean
        version:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CardProduct:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        cardType:
          type: string
          enum:
          - DEBIT
          - PREPAID
          - CREDIT
        binRangeStart:
          type: string
        binRangeEnd:
          type: string
        defaultDailyLimit:
          type: number
        features:
          type: object
          additionalProperties:
            type: object
        active:
          type: boolean
        version:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Meta:
      type: object
      properties:
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
...
    ApiResponseCard:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Card"
        meta:
          $ref: "#/components/schemas/Meta"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ApiError"
...
    IssueCardRequest:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        linkedEntityId:
          type: string
          format: uuid
        virtual:
          type: boolean
...
    SimulateRequest:
      type: object
      properties:
        cardNumber:
          type: string
        expiryDate:
          type: string
        amount:
          type: number
        currency:
          type: string
        terminalId:
          type: string
        merchantId:
          type: string
        merchantName:
          type: string
        mcc:
          type: string
        entryMode:
          type: string
        pinBlock:
          type: string
        originalStan:
          type: string
        originalRrn:
          type: string
        networkCode:
          type: string
    ApiResponseSimulateResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/SimulateResponse"
        meta:
          $ref: "#/components/schemas/Meta"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ApiError"
    SimulateResponse:
      type: object
      properties:
        requestMti:
          type: string
        responseMti:
          type: string
        responseCode:
          type: string
        responseDescription:
          type: string
        authCode:
          type: string
        availableBalance:
          type: string
        stan:
          type: string
        rrn:
          type: string
        approved:
          type: boolean
        requestHex:
          type: string
        responseHex:
          type: string
...
    ApiResponseListCard:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Card"
        meta:
          $ref: "#/components/schemas/Meta"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ApiError"
...
    ApiResponseBalanceResult:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/BalanceResult"
        meta:
          $ref: "#/components/schemas/Meta"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ApiError"
    BalanceResult:
      type: object
      properties:
        availableBalance:
          type: number
        cardType:
          type: string
...